# docs/swagger.yaml
definitions:
  controller.ArticleIn:
    properties:
      author:
        type: string
      cid:
        type: integer
      content:
        type: string
      cover:
        type: string
      title:
        type: string
    required:
    - author
    - cid
    - content
    - cover
    - title
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
info:
  contact: {}
  description: 博客主要包含用户、文章和分类
  license:
    name: Apache 2.0
  title: Gin博客项目接口文档
  version: "1.0"
paths:
  /api/article/{id}/hot:
    get:
      consumes:
      - application/json
      description: 通过文章id将文章变为热门
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 将文章变为热门
      tags:
      - 文章接口
  /api/article/{id}/off:
    get:
      consumes:
      - application/json
      description: 通过文章id将文章上架
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 下架文章
      tags:
      - 文章接口
  /api/article/{id}/on:
    get:
      consumes:
      - application/json
      description: 通过文章id将文章上架
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 上架文章
      tags:
      - 文章接口
  /api/article/category/{id}:
    get:
      consumes:
      - application/json
      description: 通过参数分类id获取该id下的文章列表
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取分类下的文章列表
      tags:
      - 文章接口
  /api/article/create:
    post:
      consumes:
      - application/json
      description: 创建文章，
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.ArticleIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建文章
      tags:
      - 文章接口
  /api/article/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 通过文章id删除指定的文章
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章title
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除文章接口
      tags:
      - 文章接口
  /api/article/detail/{id}:
    get:
      consumes:
      - application/json
      description: 通过参数文章id获取文章的详情
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取文章详情
      tags:
      - 文章接口
  /api/article/list:
    get:
      consumes:
      - application/json
      description: 默认获取全部文章列表，分页参数为page_num和page_size
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前第X页
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取文章列表
      tags:
      - 文章接口
  /api/article/search:
    get:
      consumes:
      - application/json
      description: 条件筛选文章列表，如title、status
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章名称
        in: query
        name: title
        type: string
      - description: 文章内容
        in: query
        name: content
        type: string
      - description: 文章状态
        in: query
        name: status
        type: integer
      - description: 文章热门
        in: query
        name: is_hot
        type: integer
      - description: 文章分类
        in: query
        name: cid
        type: integer
      - description: 当前第X页
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 筛选文章接口
      tags:
      - 文章接口
  /api/article/update/{id}:
    put:
      consumes:
      - application/json
      description: 通过文章id更新文章详细信息
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文章id
        in: path
        name: id
        required: true
        type: integer
      - description: 文章参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.ArticleIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新文章详情
      tags:
      - 文章接口
...
swagger: "2.0"